pyhtonScripts/requirements.txt
influxdb-client==1.38.0
paho-mqtt==1.6.1        